backend/requirements.txt
mysql-connector-python==9.4.0
python-dotenv==1.1.1
fastapi[standard]==0.116.1
pydantic==2.11.7
PyJWT==2.10.1